requirements.txt
numpy
pandas
scikit-learn
matplotlib
pathlib
pillow
ipykernel
pyyaml
setuptools
tqdm
torch
requests
fastnumbers
torchvision
opencv-python
albumentations

absl-py==1.4.0
addict==2.4.0
albumentations==1.3.1
asttokens==2.2.1
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.1
click==8.1.4
cmake==3.27.7
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.36
DateTime==5.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dpcpp-cpp-rt==2023.2.0
einops==0.6.1
executing==1.2.0
fastjsonschema==2.17.1
fastnumbers==5.0.1
filelock==3.12.4
fonttools==4.40.0
fsspec==2023.10.0
ftfy==6.1.1
gdown==4.7.1
google-auth==2.21.0
google-auth-oauthlib==1.0.0
grpcio==1.56.0
huggingface-hub==0.18.0
idna==3.4
imageio==2.31.6
importlib-metadata==6.7.0
importlib-resources==5.12.0
install==1.3.5
intel-cmplr-lib-rt==2023.2.0
intel-cmplr-lic-rt==2023.2.0
intel-opencl-rt==2023.2.0
intel-openmp==2023.2.0
ipykernel==6.26.0
ipython==8.12.3
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-pygments==0.2.2
kiwisolver==1.4.4
lazy_loader==0.3
lit==17.0.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.1
mkl==2023.2.0
mkl-fft==1.3.6
mkl-service==2.4.0
mmcv==2.0.0
mmdet==2.28.1
mmengine==0.9.0
mmsegmentation==1.0.0
model-index==0.1.11
mpmath==1.2.1
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opencv-python==4.8.0.74
opencv-python-headless==4.8.1.78
opendatalab==0.0.9
openmim==0.3.9
ordered-set==4.1.0
overrides==7.4.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.8.0
prettytable==3.9.0
prompt-toolkit==3.0.39
protobuf==4.23.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.18.0
Pygments==2.15.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.1.0
qudida==0.0.4
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.4.2
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.12.2
six==1.16.0
soupsieve==2.4.1
stack-data==0.6.2
sympy==1.12
tabulate==0.9.0
tbb==2021.10.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
terminaltables==3.1.10
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.6.11
tinycss2==1.2.1
tomli==2.0.1
torch==2.0.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
triton==2.0.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==1.26.18
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
yacs==0.1.8
yapf==0.40.1
zipp==3.15.0
zope.interface==6.0